Improve linux system identification
RbConfig will report some Linux systems as 'linux-eabi' or similar,
depending on architectue. By checking the target-os stats with 'linux',
rather than using direct equality, we make this check more accurate. I
investigated this as scout_realtime was not showing a network graph on
my Raspberry Pi, as it was not correctly identifying it as a Linux
system and was therefore skipping that collector.